chore: refactored as class
